Updates
Major overhaul to the script, adding automatic checks for Windows updates,
as well as prompts for the user to enter their desired server name/password/port numbers.  Additionally, added a function to create firewall rules based on the port numbers chosen.  Lastly, added a system check to check in a system reboot is needed and prompt the user.  If the server requires a reboot, the user can choose to reboot or not.